+++ hello-world/package.json
@@ -5,17 +5,21 @@
   "scripts": {
     "build": "node ace build --production",
     "start": "node server.js",
+    "test": "jest",
     "dev": "node ace serve --watch",
     "lint": "eslint . --ext=.ts"
   },
   "devDependencies": {
     "@adonisjs/assembler": "^5.3.2",
+    "@types/jest": "^26.0.23",
     "adonis-preset-ts": "^2.1.0",
     "eslint": "^7.29.0",
     "eslint-plugin-adonis": "^1.3.2",
     "jest": "^27.0.6",
     "pino-pretty": "^5.1.0",
-    "typescript": "^4.2.4",
+    "supertest": "^6.1.3",
+    "ts-jest": "^27.0.3",
+    "typescript": "^4.3.5",
     "youch": "^2.2.2",
     "youch-terminal": "^1.1.1"
   },
@@ -25,5 +29,22 @@
     "proxy-addr": "^2.0.7",
     "reflect-metadata": "^0.1.13",
     "source-map-support": "^0.5.19"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.ts$": "ts-jest"
+    },
+    "testRegex": "\\.test\\.ts$",
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "json"
+    ],
+    "globals": {
+      "ts-jest": {
+        "diagnostics": true
+      }
+    }
   }
 }
