+++ hello-world/start/routes.ts
@@ -19,68 +19,20 @@
 */
 
 import Route from '@ioc:Adonis/Core/Route'
+const userController = require('../controllers/user')
 
 // Route.get('/', async () => {
 //   return 'hello world'
 // })
 
-class User {
-  public index: number;
-  public id: string;
-
-  constructor(
-    index: number,
-    id: string,
-  ) {
-    this.index = index;
-    this.id = id;
-  }
-}
-
-const user1: User = new User(1, 'user1');
-const user2: User = new User(2, 'user2');
-
-let userDB: User[] = [user1, user2];
-let getnextIdx = (userDB: User[]): number => {
-  return userDB.length +1;
-} 
-
 // 모든 유저 조회 (Read)
-Route.get('/', ({ response }) => {
-  response.send(userDB);
-})
+Route.get('/', userController.read)
 
 // 유저 추가 (Create)
-Route.post('/', ({ request, response }) => {
-  const newUser = request.body().id;
-  if (!newUser) {
-    return response.status(400).send({ 'message': 'required id' });
-  }
-  userDB.push(new User(getnextIdx(userDB), newUser));
-  response.send(userDB[getnextIdx(userDB) -1]);
-})
-// 유저 중복아이디 있는지 확인
+Route.post('/', userController.create)
 
 // 유저 아이디 변경 (Update)
-Route.patch('/:index', ({ request, params , response}) => {
-  const userId = request.body().id;
-  const index = params.index -1;
-  if (!userId) {
-    return response.status(400).send({ 'message': 'required id' });
-  }
-  userDB[index]['id'] = userId;
-  response.send(`changed as ${userDB[index]['id']}`);
-})
+Route.patch('/:index', userController.update)
 
 // 유저 삭제 (Delete)
-Route.delete('/:index', ({ params, response }) => {
-  const index = params.index;
-  if (!index) {
-    return response.status(400).send({ 'message': 'required index' });
-  }
-  if (Number(index) > userDB.length) {
-    return response.status(404).send({ 'message': `user index : ${index} not found`});
-  }
-  userDB = userDB.filter((user) => user.index !== Number(index));
-  response.send({ 'message' : `deleted index : ${index}` });
-})
+Route.delete('/:index', userController.delete)
